Move FilterBadge in its own file. Add click handler.
When clicking on the badge, it should open the corresponding filter
dialog for editing the filter.
